refactor: isolate forgot password
