+++ backend/src/main/kotlin/be/osoc/team1/backend/controllers/UserController.kt
@@ -78,21 +78,4 @@ class UserController(private val service: UserService) {
     @Secured("ROLE_ADMIN")
     fun postUserRole(@PathVariable id: UUID, @RequestBody role: Role) =
         service.changeRole(id, role)
-
-    /**
-     * Request to reset the password of the user with given [emailAddress]. The link to actually reset the password is
-     * sent in an email to [emailAddress].
-     * This request will always succeed, even when an invalid [emailAddress] is given. Otherwise, people with bad intent
-     * could track down which email addresses are linked to existing accounts.
-     */
-    @PostMapping("/resetPassword")
-    @ResponseStatus(value = HttpStatus.NO_CONTENT)
-    fun postEmail(@RequestBody emailAddress: String) = service.sendEmailWithToken(emailAddress)
-
-    /**
-     * Reset password using [resetPasswordUUID].
-     */
-    @PatchMapping("/resetPassword/{resetPasswordUUID}")
-    fun patchPassword(@PathVariable resetPasswordUUID: UUID, @RequestBody newPassword: String) =
-        service.changePassword(resetPasswordUUID, newPassword)
 }
